// src/types/studentStates.types.ts
export enum StudentStates {
    PENDING = "pending",
    IN_PROGRESS = "in_progress",
    TIME_OUT = "time_out",
    TO_CALL = "to_call",
    CALLED = "called",
    TO_REGISTERED = "to_registered",
    REGISTERED = "registered",
    NO_REGISTERED = "no_registered",
    CHECKED = "checked",
    DONE = "done",
}